requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
pypdf==6.5.0
PyJWT==2.8.0
reportlab==4.0.7
#VERCION PYTHON 3.12.10